assets/training/general/environments/tensorflow-2.16-cuda11/tests/requirements.txt
azure-ai-ml==1.30.0
azure.identity==1.10.0
